requirements.txt
fastapi
uvicorn[standard]
requests
openai
python-dotenv
cachetools